+++ gbhw.h
@@ -1,4 +1,4 @@
-/* $Id: gbhw.h,v 1.13 2004/10/23 21:12:49 ranmachan Exp $
+/* $Id: gbhw.h,v 1.14 2005/06/29 00:34:57 ranmachan Exp $
  *
  * gbsplay is a Gameboy sound player
  *
@@ -16,29 +16,29 @@
 
 struct gbhw_buffer {
 	int16_t *data;
-	int pos;
-	int len;
+	long pos;
+	long len;
 };
 
 struct gbhw_channel {
-	int mute;
-	int master;
-	int leftgate;
-	int rightgate;
-	int volume;
-	int env_dir;
-	int env_tc;
-	int env_ctr;
-	int sweep_dir;
-	int sweep_tc;
-	int sweep_ctr;
-	int sweep_shift;
-	int len;
-	int len_enable;
-	int div_tc;
-	int div_ctr;
-	int duty_tc;
-	int duty_ctr;
+	long mute;
+	long master;
+	long leftgate;
+	long rightgate;
+	long volume;
+	long env_dir;
+	long env_tc;
+	long env_ctr;
+	long sweep_dir;
+	long sweep_tc;
+	long sweep_ctr;
+	long sweep_shift;
+	long len;
+	long len_enable;
+	long div_tc;
+	long div_ctr;
+	long duty_tc;
+	long duty_ctr;
 };
 
 extern struct gbhw_channel gbhw_ch[4];
@@ -46,12 +46,12 @@ extern struct gbhw_channel gbhw_ch[4];
 typedef regparm void (*gbhw_callback_fn)(struct gbhw_buffer *buf, void *priv);
 
 regparm void gbhw_setcallback(gbhw_callback_fn fn, void *priv);
-regparm void gbhw_setrate(int rate);
+regparm void gbhw_setrate(long rate);
 regparm void gbhw_setbuffer(struct gbhw_buffer *buffer);
 regparm void gbhw_init(uint8_t *rombuf, uint32_t size);
-regparm void gbhw_pause(int new_pause);
-regparm void gbhw_master_fade(int speed, int dstvol);
+regparm void gbhw_pause(long new_pause);
+regparm void gbhw_master_fade(long speed, long dstvol);
 regparm void gbhw_getminmax(int16_t *lmin, int16_t *lmax, int16_t *rmin, int16_t *rmax);
-regparm int gbhw_step(int time_to_work);
+regparm long gbhw_step(long time_to_work);
 
 #endif
